+++ pandas/tests/indexes/test_index_new.py
@@ -358,6 +358,15 @@ class TestDtypeEnforced:
 
         tm.assert_numpy_array_equal(idx.values, expected.values)
 
+    def test_pass_datetimeindex_to_index(self):
+        # GH#1396
+        rng = date_range("1/1/2000", "3/1/2000")
+        idx = Index(rng, dtype=object)
+
+        expected = Index(rng.to_pydatetime(), dtype=object)
+
+        tm.assert_numpy_array_equal(idx.values, expected.values)
+
 
 class TestIndexConstructorUnwrapping:
     # Test passing different arraylike values to pd.Index
